+++ auth-service/src/main/java/com/devspark/authservice/AuthServiceApplication.java
@@ -1,0 +1,13 @@
+package com.devspark.authservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AuthServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(AuthServiceApplication.class, args);
+	}
+
+}
