+++ auth-service/src/main/java/com/devspark/authservice/constants/DeleteFlags.java
@@ -1,0 +1,9 @@
+package com.devspark.authservice.constants;
+
+/**
+ * constant represent delete flags
+ */
+public record DeleteFlags() {
+    public static final int NOT_DELETED = 0;
+    public static final int DELETED = 1;
+}
