+++ auth-service/src/main/java/com/devspark/authservice/exception/ErrorDetails.java
@@ -1,0 +1,4 @@
+package com.devspark.authservice.exception;
+
+public record ErrorDetails(String timestamp, String message, String path, String errorCode) {
+}
