+++ auth-service/src/main/java/com/devspark/authservice/exception/customExceptions/IncorrectPasswordException.java
@@ -1,0 +1,13 @@
+package com.devspark.authservice.exception.customExceptions;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.BAD_REQUEST)
+public class IncorrectPasswordException extends RuntimeException {
+    private String message;
+
+    public IncorrectPasswordException(String message) {
+        super(message);
+    }
+}
