+++ auth-service/src/main/java/com/devspark/authservice/pojo/dto/LoginUserDTO.java
@@ -1,0 +1,6 @@
+package com.devspark.authservice.pojo.dto;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record LoginUserDTO(@NotBlank(message = "username cannot be null") String username, @NotBlank(message = "password cannot be null")String password) {
+}
