+++ auth-service/src/main/java/com/devspark/authservice/pojo/vo/CreateUserVO.java
@@ -1,0 +1,4 @@
+package com.devspark.authservice.pojo.vo;
+
+public record CreateUserVO(Boolean success) {
+}
