+++ auth-service/src/main/java/com/devspark/authservice/repository/AuthRepository.java
@@ -1,0 +1,11 @@
+package com.devspark.authservice.repository;
+
+import com.devspark.authservice.entity.UserInfoEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface AuthRepository extends JpaRepository<UserInfoEntity, Long> {
+
+    Optional<UserInfoEntity> getUserInfoEntityByUsername(String username);
+}
