+++ auth-service/src/main/resources/cert/private.pem
