+++ auth-service/src/main/resources/cert/public.pem
