+++ auth-service/src/test/java/com/devspark/authservice/UserserviceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.devspark.authservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserserviceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
