--- user-service/src/main/java/com/devspark/userservice/constants/DeleteFlags.java
@@ -1,9 +1,0 @@
-package com.devspark.userservice.constants;
-
-/**
- * constant represent delete flags
- */
-public record DeleteFlags() {
-    public static final int NOT_DELETED = 0;
-    public static final int DELETED = 1;
-}
