--- user-service/src/main/java/com/devspark/userservice/exception/ErrorDetails.java
@@ -1,4 +1,0 @@
-package com.devspark.userservice.exception;
-
-public record ErrorDetails(String timestamp, String message, String path, String errorCode) {
-}
