--- user-service/src/main/java/com/devspark/userservice/exception/customExceptions/JWTSignException.java
@@ -1,13 +1,0 @@
-package com.devspark.userservice.exception.customExceptions;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
-public class JWTSignException extends RuntimeException {
-    private String message;
-
-    public JWTSignException(String message) {
-        super(message);
-    }
-}
