--- user-service/src/main/java/com/devspark/userservice/exception/customExceptions/UserNotFoundException.java
@@ -1,13 +1,0 @@
-package com.devspark.userservice.exception.customExceptions;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.BAD_REQUEST)
-public class UserNotFoundException extends RuntimeException{
-    private String message;
-
-    public UserNotFoundException(String message) {
-        super(message);
-    }
-}
