--- user-service/src/main/java/com/devspark/userservice/pojo/vo/CreateUserVO.java
@@ -1,4 +1,0 @@
-package com.devspark.userservice.pojo.vo;
-
-public record CreateUserVO(Boolean success) {
-}
