--- user-service/src/main/java/com/devspark/userservice/pojo/vo/LoginUserVO.java
@@ -1,4 +1,0 @@
-package com.devspark.userservice.pojo.vo;
-
-public record LoginUserVO(String token, String statusCode) {
-}
