--- user-service/src/main/resources/application.properties
@@ -1,25 +1,0 @@
-# application
-server.port=8081
-spring.application.name=micro-user-auth
-
-#spring.datasource.url=jdbc:postgresql://postgre16:5432/db_user_service
-spring.datasource.url=jdbc:postgresql://localhost:5432/db_user_service
-spring.datasource.username=root
-spring.datasource.password=secret
-
-# jpa configuration
-#spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
-
-# logging configuration
-logging.level.root=WARN
-logging.level.com.example.springscheduler = debug
-logging.level.org.springframework.web = WARN
-logging.file.name=log/devspark.log
-logging.logback.rollingpolicy.max-file-size=10MB
-logging.logback.rollingpolicy.max-history=10
-logging.pattern.console=%d{yyyy-MM-dd HH:mm:ss} - %logger{36} - %msg%n
-
-# jwt
-rsa.private-key=MIIEvgIBADANBgkqhkiG9w0BAQEFAASCBKgwggSkAgEAAoIBAQCoFmqVleuyCqn942XwRXIaQkar8QTUrn8F1yhyq2+yZNNf0nv7FAhhvCgCqssoF/3+l55KdPW3zStimQxF4tRcC7xp8eUb6/sAMyU3jIbjof6DHSCt1pgCGUkXJFqEPL4UIsMerhFo4BZLYB9JyFIErCymckx1q0C3yEbeCXnPRcOjJsaNWtH6HdAgIvSR1srHnSn1CbpJ7UeOeyrfsUD31ssr82cFiGuPO6om1H1Ds6c0/R5sBaUxWxggvRDWKeTEwT6eK/CHEIObd0s6CmS5R7Uuk9yUaFfZzrDkxqwmEzh7mGn/MlAWMMsyPFqrD/pPsyWkvjkKK6CsqPn9tt9jAgMBAAECggEAExjo5zHvY30erz/k1xWszxr0mJh7qgZhXDKD3/Nmb2hRM7Pgbjhqc2BDU3pANr/zFbf4PhFn0Yg4A63L3SzWX9bdg1P+adu32t/0vIj00mwXVKI8kkOtZZ+l1EFHH3sfRX5lmNjCFZ5QKiQzqdheqTU+pXKbCUt5wOe6VYPSGSBu2iVUhX4XMxdqsManqkxsO8tAJMM03q8gsjd4tT2J1ZiHxG8FHr0L2XmcT/+Yafrj+HXyyPEQSbOpepMATGlPKbnrCl8t+buu017pQBXpiV0XrUs1RTPeKbjZ6VR5RPfUaFBpUdjtVIgT1gVym0USVRtz5EOeQlSzJUPR91UGoQKBgQDYWRFcVHiis7zVlGcqG6mSuzvQI5Ry7PvAZRPTafbKij7EYRb8yd1BH0j/bw6kA/rClQX4OJGnlBCdvsik1fqXmM5wyYPTqk4ZCLa1HPVeQoGrIy/VOx5dTa0R/JRUb3sTEmDcTSThxPR30nZJTiPqtg7EeI9SGZH2ps0atPORmwKBgQDG5P9jYvHWq0zFLhpBHJO5c1WYQN/fQvOctOpep8dIuzac71iuy/5qgVz61rgDahp3uoKfvbz5L/YitAoweSZyIUDSfxwIejxPyeiD/rfdtpAYIJuRnVLwhbOdUbuSCtvCqgf7ovwzM2nVVJcU+5EIt8RjYfjExFp9d2uKvIYJ2QKBgQCs5sHRIFEo9qw7pNat+F2VPr2WoljJYEvyftVRaKYrSEDRgSZ18H46xV35hjAVNHj7eU3ECrGdU9/Tsd14nuoeraiOyv3n83tm1mfuZq5SHqSyuPRJXW3u810Vut5wfIfHvjBa3+xR4q3d9P1DqIdb1RQIQLGJ05WrobLZXTQrcQKBgAHn59qX9CFOeunYyz+cen3ocQWIeZy2E4eI5EYL7Yjw4lEWqg7ivYlfu/hbXmAKAD7tdpr5QnpQz1b9yqs0Pf4R+4EI3sfmA/XvxV6ebjydpzvT9CmtDEJpxZLYWafImtp+rpzmruv+riDxMpFs57Lhr+wT4EJib0uMfWdS7yXhAoGBAKNONA7xczKQTsBIfh8cn5yfio9d48M73aXYrDEGTuIFWnaaJ8d3t7I3OzXX9BApbYBQAntb4rZcb1Z8+2ICC/OiciirCO0Ch6qF5JzBUFvyqqGfn47iOWc4eKge81QtGuZoOh4A4b0b8dkWUsEA4dfOqVFSPDBZtrx4tihKMah2
-rsa.public-key=MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAqBZqlZXrsgqp/eNl8EVyGkJGq/EE1K5/BdcocqtvsmTTX9J7+xQIYbwoAqrLKBf9/peeSnT1t80rYpkMReLUXAu8afHlG+v7ADMlN4yG46H+gx0grdaYAhlJFyRahDy+FCLDHq4RaOAWS2AfSchSBKwspnJMdatAt8hG3gl5z0XDoybGjVrR+h3QICL0kdbKx50p9Qm6Se1Hjnsq37FA99bLK/NnBYhrjzuqJtR9Q7OnNP0ebAWlMVsYIL0Q1inkxME+nivwhxCDm3dLOgpkuUe1LpPclGhX2c6w5MasJhM4e5hp/zJQFjDLMjxaqw/6T7MlpL45CiugrKj5/bbfYwIDAQAB
