--- user-service/src/test/java/com/devspark/userservice/UserserviceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.devspark.userservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class UserserviceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
